+++ changelog.md
@@ -2,6 +2,8 @@
 -----
 
 - Fix mentioning to retweeted status.
+- Add streaming mode on after_death command.
+- Filter tweet by myself before get answer.
 
 
 0.2.1